@@ -5,36 +5,41 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "deploy": "ng build --prod && firebase deploy --only hosting",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/cdk": "~7.0.2",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
+    "@angular/animations": "~7.1.3",
+    "@angular/cdk": "~7.1.1",
+    "@angular/common": "~7.1.3",
+    "@angular/compiler": "~7.1.3",
+    "@angular/core": "~7.1.3",
     "@angular/fire": "^5.1.0",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/material": "^7.0.2",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/forms": "~7.1.3",
+    "@angular/http": "~7.1.3",
+    "@angular/material": "^7.1.1",
+    "@angular/platform-browser": "~7.1.3",
+    "@angular/platform-browser-dynamic": "~7.1.3",
+    "@angular/pwa": "^0.11.3",
+    "@angular/router": "~7.1.3",
+    "@angular/service-worker": "~7.1.3",
     "core-js": "^2.5.4",
     "firebase": "^5.5.7",
+    "firebase-tools": "^6.1.2",
     "hammerjs": "^2.0.8",
+    "npm": "^6.5.0",
     "rxjs": "~6.3.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.3",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
+    "@angular/cli": "~7.1.3",
+    "@angular/compiler-cli": "~7.1.3",
+    "@angular/language-service": "~7.1.3",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
