@@ -3,7 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const functions = require("firebase-functions");
 const admin = require("firebase-admin");
 admin.initializeApp();
-const db = admin.firestore();
+const firestore = admin.firestore();
 exports.addMessage = functions.https.onRequest((req, res) => {
     // Grab the text parameter.
     const original = req.query.text;
@@ -13,6 +13,31 @@ exports.addMessage = functions.https.onRequest((req, res) => {
         return res.redirect(303, snapshot.ref.toString());
     });
 });
+exports.deleteOldPartiesAndGameInstances = functions.firestore
+    .document('parties/{partyId}')
+    .onCreate((snap, context) => {
+    const now = Date.now();
+    const cutoff = now - 60 * 60 * 24 * 1000; // deleting 1 day old parties
+    const promises = [];
+    promises.push(deleteCollection(firestore, 'parties', cutoff));
+    promises.push(deleteCollection(firestore, 'game-instances', cutoff));
+    return Promise.all(promises);
+});
+function deleteCollection(db, collectionPath, cutoff) {
+    const promises = [];
+    return db.collection(collectionPath).where('created', '<=', cutoff)
+        .get()
+        .then(snapshot => {
+        console.log('Deleting old ' + collectionPath);
+        snapshot.forEach(doc => {
+            promises.push(doc.ref.delete());
+        });
+        return Promise.all(promises);
+    }).catch(err => {
+        console.log('Something went wrong when deleting', err);
+        return false;
+    });
+}
 /*// Nihe aggregate number of questions
 exports.aggregateNhieQuestions = functions.firestore
   .document('nhie-questions/{nId}')
