@@ -1,8 +1,8 @@
 import * as functions from 'firebase-functions';
-import * as admin from 'firebase-admin'
+import * as admin from 'firebase-admin';
 admin.initializeApp();
-const db = admin.firestore();
 
+const firestore = admin.firestore();
 
 
 exports.addMessage = functions.https.onRequest((req, res) => {
@@ -15,8 +15,37 @@ exports.addMessage = functions.https.onRequest((req, res) => {
   });
 });
 
+exports.deleteOldPartiesAndGameInstances = functions.firestore
+  .document('parties/{partyId}')
+  .onCreate((snap, context) => {
 
+  const now = Date.now();
+  const cutoff = now - 60 * 60 * 24 * 1000; // deleting 1 day old parties
 
+  const promises = [];
+
+  promises.push(deleteCollection(firestore,'parties', cutoff));
+  promises.push(deleteCollection(firestore,'game-instances', cutoff));
+
+  return Promise.all(promises);
+});
+
+function deleteCollection(db, collectionPath, cutoff) {
+  const promises = [];
+
+  return db.collection(collectionPath).where('created', '<=', cutoff)
+    .get()
+    .then(snapshot => {
+      console.log('Deleting old ' + collectionPath);
+      snapshot.forEach(doc => {
+        promises.push(doc.ref.delete());
+      });
+      return Promise.all(promises);
+    }).catch(err => {
+      console.log('Something went wrong when deleting', err);
+      return false;
+    });
+}
 
 /*// Nihe aggregate number of questions
 exports.aggregateNhieQuestions = functions.firestore
