@@ -2,15 +2,33 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { NhieComponent } from './games/nhie/nhie.component';
 import { PageNotFoundComponent } from './shared/page-not-found/page-not-found.component';
-import { GamesComponent} from './games/games.component';
-import { LobbyComponent} from './party/lobby/lobby.component';
+import { LobbyComponent } from './party/lobby/lobby.component';
 import { AliasComponent } from './party/alias/alias.component';
-import {GameInfoComponent} from './party/game-info/game-info.component';
+import { GameInfoComponent } from './party/game-info/game-info.component';
+import { HomeComponent } from './party/home/home.component';
+import { VotingGameComponent } from './games/voting-game/voting-game.component';
+import { WrapperComponent } from './party/wrapper/wrapper.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: GamesComponent
+    component: HomeComponent,
+    pathMatch: 'full'
+  },
+  {
+    path: '',
+    component: WrapperComponent,
+    children: [
+      { path: 'alias', component: AliasComponent },
+      { path: 'alias/:joinCode', component: AliasComponent },
+      { path: 'alias/:joinCode/:fromLink', component: AliasComponent },
+      { path: 'lobby/:joinCode', component: LobbyComponent },
+      { path: 'game-info/:gameId', component: GameInfoComponent },
+      { path: 'game/nhie/:joinCode', component: NhieComponent },
+      { path: 'game/nhie', component: NhieComponent },
+      { path: 'voting/:joinCode', component: VotingGameComponent },
+      { path: 'voting', component: VotingGameComponent },
+    ]
   },
   {
     path: 'alias',
@@ -18,6 +36,10 @@ const routes: Routes = [
   },
   {
     path: 'alias/:joinCode',
+    component: AliasComponent
+  },
+  {
+    path: 'alias/:joinCode/:fromLink',
     component: AliasComponent
   },
   {
@@ -43,7 +65,7 @@ const routes: Routes = [
   {
     path: '**',
     component: PageNotFoundComponent,
-  }
+  },
 ];
 
 @NgModule({
