@@ -1,4 +1,5 @@
 import {Component, EventEmitter, OnInit, Output} from '@angular/core';
+import {MenuService} from '../../../party/shared/menu.service';
 
 @Component({
   selector: 'app-add-manual-question',
@@ -11,13 +12,20 @@ export class AddManualQuestionComponent implements OnInit {
 
   @Output() submitNewQuestion = new EventEmitter<string>();
 
-  constructor() { }
+  @Output() hideAddQuestion = new EventEmitter<string>();
+
+  constructor(private menuService: MenuService) { }
 
   ngOnInit() {
   }
 
   submitNewManualQuestionClick() {
+    this.menuService.setHideAll(false);
     this.submitNewQuestion.emit(this.newManualQuestion);
   }
 
+  hideQuestionClick() {
+    this.menuService.setHideAll(false);
+    this.hideAddQuestion.emit('close');
+  }
 }
