@@ -1,4 +1,4 @@
-import {Component, Input} from '@angular/core';
+import {Component, ElementRef, Input, ViewChild} from '@angular/core';
 
 @Component({
   selector: 'app-nhie-card',
@@ -7,8 +7,21 @@ import {Component, Input} from '@angular/core';
 })
 export class NhieCardComponent {
 
+  private _question: string;
+
   @Input()
-  question: string;
+  set question(question: string) {
+    console.log('prev value: ', this._question);
+    console.log('got name: ', question);
+    this.isLoading = true;
+    setTimeout(() => {
+      this.isLoading = false;
+    }, 400);
+
+    this._question = question;
+  }
+
+  isLoading: boolean;
 
   constructor() {}
 
