@@ -1,6 +1,7 @@
 export interface GameInstance {
   id?: string;
   partyId: string;
-  joinCode: string;
+  joinCode: number;
   gameId: string;
+  created: number;
 }
