@@ -1,11 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {Location} from '@angular/common';
 import {ActivatedRoute} from '@angular/router';
 import {GameService} from '../../games/shared/game.service';
 import {Observable} from 'rxjs';
 import {Game} from '../../games/shared/game.model';
-import {Party} from '../shared/party';
-import {PartyService} from '../shared/party.service';
+
+import {MenuService} from '../shared/menu.service';
 
 @Component({
   selector: 'app-game-info',
@@ -18,29 +18,15 @@ export class GameInfoComponent implements OnInit {
 
   gameObs: Observable<Game>;
 
-  joinCode: string;
-
-  image_path: string;
-
   constructor(private route: ActivatedRoute,
               private gameService: GameService,
-              private location: Location,
-              private partyService: PartyService) { }
+              private menuService: MenuService) { }
 
   ngOnInit() {
+    this.menuService.setMenuVisibility(false);
+    this.menuService.setRouterlink(null);
     this.gameId = this.route.snapshot.params['gameId'];
     this.gameObs = this.gameService.getGame(this.gameId);
-    this.gameObs.subscribe(data => {
-      this.image_path = data.image_path;
-    });
-
-    this.partyService.party.subscribe((party: Party) => {
-      this.joinCode = party.joinCode;
-    });
-  }
-
-  navigateBack() {
-    this.location.back();
   }
 
 }
