@@ -1,14 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import { Party } from '../shared/party';
 import { PartyService } from '../shared/party.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { GameInstanceService } from '../../games/shared/game-instance.service';
 import { NhieGameInstanceService } from '../../games/nhie/shared/nhie-game-instance.service';
-import { combineLatest, Observable } from 'rxjs';
+import {combineLatest, Observable, Subscription} from 'rxjs';
 import {GameService} from '../../games/shared/game.service';
 import {Game} from '../../games/shared/game.model';
 import {AuthService} from '../../core/auth/auth.service';
 import { tap, take } from 'rxjs/operators';
+import {MenuService} from '../shared/menu.service';
 
 
 @Component({
@@ -16,41 +17,38 @@ import { tap, take } from 'rxjs/operators';
   templateUrl: './lobby.component.html',
   styleUrls: ['./lobby.component.scss']
 })
-export class LobbyComponent implements OnInit {
+export class LobbyComponent implements OnInit, OnDestroy {
 
-  party: Observable<Party>;
-  gameObservable: Observable<Game>;
   waitingMessage = 'Waiting for Game Leader <br> to start the game...';
+  party: Observable<Party>;
   isLoggedIn: Observable<boolean>;
-
   isGameLeader: Observable<boolean>;
 
-  joinCode: string;
+  gameInstanceSub: Subscription;
+
 
   constructor(
     private partyService: PartyService,
     private router: Router,
     private gameInstanceService: GameInstanceService,
-    private nhieService: NhieGameInstanceService,
     private route: ActivatedRoute,
     private gameService: GameService,
-    private authService: AuthService
+    private authService: AuthService,
+    private menuServie: MenuService
   ) { }
 
   ngOnInit() {
+    this.menuServie.setMenuVisibility(true);
     this.party = this.partyService.party;
     this.isGameLeader = this.partyService.isGameLeaderObservable;
     this.isLoggedIn = this.authService.isLoggedInObservable;
-    this.joinCode = this.route.snapshot.params['joinCode'];
-    this.joinCode = this.joinCode.toLowerCase();
-    this.gameObservable = this.gameService.game;
     this.checkGameInstance();
   }
 
   checkGameInstance() {
-    this.gameInstanceService.gameInstance.subscribe(gameInstance => {
+    this.gameInstanceSub = this.gameInstanceService.gameInstance.subscribe(gameInstance => {
       if (gameInstance && this.router.url.indexOf('lobby') !== -1) {
-        this.router.navigate([gameInstance.gameId + '/' + gameInstance.joinCode]);
+        this.router.navigate(['game/' + gameInstance.gameId + '/' + gameInstance.joinCode]);
       }
     });
   }
@@ -59,19 +57,14 @@ export class LobbyComponent implements OnInit {
     this.party.pipe(
       take(1)
     ).subscribe(party => {
-      this.router.navigate([party.selectedGame + '/' + party.joinCode]);
+      this.router.navigate(['game/' + party.selectedGame + '/' + party.joinCode]);
     });
   }
 
-  onGameInfoClick() {
-    this.gameObservable.pipe(
-      take(1)
-    ).subscribe(game => {
-      this.router.navigate(['game-info/' + game.urlPath]);
-    });
-  }
-
-  onLeaveClick() {
+  ngOnDestroy(): void {
+    if (this.gameInstanceSub) {
+      this.gameInstanceSub.unsubscribe();
+    }
 
   }
 }
