@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { DefaultSizeButtonComponent } from './default-size-button/default-size-button.component';
-import {MatIconModule, MatButtonModule, MatFormFieldModule, MatInputModule, MatMenuModule} from '@angular/material';
+import { MatIconModule, MatButtonModule, MatFormFieldModule, MatInputModule, MatMenuModule,
+   MatTooltipModule, MatProgressSpinnerModule } from '@angular/material';
 import { FooterImageComponent } from './footer-image/footer-image.component';
 import { RouterModule } from '@angular/router';
 import { BackButtonComponent } from './back-button/back-button.component';
@@ -10,6 +11,9 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { GameCodeComponent } from './game-code/game-code.component';
 import { MenuButtonComponent } from './menu-button/menu-button.component';
 import { BackgroundColourDirective } from './background-colour.directive';
+import { TruncatePipe } from './truncate-pipe/truncate.pipe';
+import { SpinnerComponent } from './spinner/spinner.component';
+import { CoreModule } from '../core/core.module';
 
 @NgModule({
   declarations: [
@@ -20,7 +24,9 @@ import { BackgroundColourDirective } from './background-colour.directive';
     GameCodeComponent,
     MenuButtonComponent,
     BackButtonComponent,
-    BackgroundColourDirective
+    BackgroundColourDirective,
+    TruncatePipe,
+    SpinnerComponent
   ],
   imports: [
     CommonModule,
@@ -31,7 +37,10 @@ import { BackgroundColourDirective } from './background-colour.directive';
     MatFormFieldModule,
     MatMenuModule,
     MatInputModule,
-    ReactiveFormsModule
+    MatTooltipModule,
+    ReactiveFormsModule,
+    MatProgressSpinnerModule,
+    CoreModule
   ],
   exports: [
     FooterImageComponent,
@@ -39,7 +48,9 @@ import { BackgroundColourDirective } from './background-colour.directive';
     DefaultSizeButtonComponent,
     GameCodeComponent,
     MenuButtonComponent,
-    BackgroundColourDirective
+    BackgroundColourDirective,
+    TruncatePipe,
+    SpinnerComponent
   ]
 })
 export class SharedModule { }
